argh finished

+++ tf_idf_lib/include/hashtable.h
@@ -11,7 +11,7 @@
 typedef struct hash_item_t {
   
   char* key;
-  unsigned int counter;
+  float counter;
 } hash_item_t;
 
 typedef struct hashtable_t {
